requirements.txt
torchio>=0.18.0
antspyx>=0.3.0
nibabel>=3.0.0
numpy>=1.19.0
matplotlib>=3.5.0
opencv-python>=4.5.0
